Trying to implement onenter and onexit
what I've done so far really doesn't make sense but I was confused as to how to actually use OnEnter/OnExit / when to impelement them, and how to tell them to be executed in terms of commands. Lulu had technical troubles and her zoom shut down. We will discuss / work on this more next class